requirements.txt
altgraph==0.17.4
chardet==5.2.0
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
reportlab==4.2.5
setuptools==75.6.0
six==1.16.0
